refactor: Split business logic in RegisterPage
