+++ frontend/src/pages/register/RegisterPage.tsx
@@ -1,8 +1,3 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { useUserAuthorizedMutation } from '../../features';
 import {
   Background,
   ChildTemplate,
@@ -10,48 +5,12 @@ import {
   ParentTemplate,
   RegisterForm,
   UserForm,
-  registerSchema,
 } from '../../components';
-
-export type RegisterData = {
-  userEmail: string;
-  userPassword: string;
-  userName: string;
-  userConfirmPassword: string;
-};
+import { useRegisterPage } from './RegisterPage.hook';
 
 export default function RegisterPage() {
-  const navigate = useNavigate();
-
-  // Register via api
-  const [userAuthorize, { error: registerError, data }] =
-    useUserAuthorizedMutation();
-
-  // Submit register
-  const registerSubmit: SubmitHandler<RegisterData> = (data: RegisterData) => {
-    const user = {
-      name: data.userName,
-      email: data.userEmail,
-      password: data.userPassword,
-    };
-    userAuthorize(user);
-  };
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors: inputErrors },
-  } = useForm<RegisterData>({
-    mode: 'onChange',
-    resolver: yupResolver(registerSchema),
-  });
-
-  // Go home after register
-  useEffect(() => {
-    if (data) {
-      navigate('/');
-    }
-  }, [data]);
+  const [register, handleSubmit, inputErrors, registerSubmit, registerError] =
+    useRegisterPage();
 
   return (
     <Layout>
